requirements.txt
fastapi
uvicorn
google-generativeai
langchain
python-dotenv
pandas
